+++ repology_client/constants.py
@@ -6,29 +6,29 @@ Hardcoded constants for rapidly-changing Repology API. What could possibly go
 wrong?
 """
 
-# Library package name.
+#: Library package name.
 PACKAGE = "repology-client"
 
-# Library version.
+#: Library version.
 VERSION = "0.1.0"
 
-# Library homepage.
+#: Library homepage.
 HOMEPAGE = "https://repology-client.sysrq.in"
 
-# Library's User-agent header
+#: Library's User-agent header
 USER_AGENT = f"Mozilla/5.0 (compatible; {PACKAGE}/{VERSION}; +{HOMEPAGE})"
 
-# Base URL for API v1 requests.
+#: Base URL for API v1 requests.
 API_V1_URL = "https://repology.org/api/v1"
 
-# Base URL for Experimental API requests.
+#: Base URL for Experimental API requests.
 API_EXP_URL = "https://repology.org/api/experimental"
 
-# Base URL for the "Project by package name" tool.
+#: Base URL for the "Project by package name" tool.
 TOOL_PROJECT_BY_URL = "https://repology.org/tools/project-by"
 
-# Maximum number of projects API can return.
+#: Maximum number of projects API can return.
 MAX_PROJECTS = 200
 
-# Number of projects, starting from which you should use bulk export instead.
+#: Number of projects, starting from which you should use bulk export instead.
 HARD_LIMIT = 5_000